+++ src/applications/find-va-forms/components/TopTasks.jsx
@@ -1,52 +1,6 @@
 import React from 'react';
-import classnames from 'classnames';
-
-function Card({ heading, href, description }) {
-  const containerClassName = classnames(
-    'vads-u-display--block',
-    'vads-u-background-color--primary-alt-lightest',
-    'vads-u-padding--2',
-    'vads-u-height--full',
-    'vads-u-text-decoration--none',
-  );
-
-  return (
-    <a href={href} className={containerClassName}>
-      <div className="vads-u-font-family--serif vads-u-font-weight--bold vads-u-text-decoration--underline">
-        {heading}
-      </div>
-      <div className="vads-l-gird-container">
-        <div className="vads-l-row">
-          <div className="vads-l-col--3 vads-u-margin-y--1">
-            <hr className="vads-u-margin--0 vads-u-border-color--cool-blue-lighter" />
-          </div>
-        </div>
-      </div>
-      <p className="vads-u-color--base">{description}</p>
-    </a>
-  );
-}
-
-function CardRow({ children: cards }) {
-  const columnClassName = classnames(
-    'vads-l-col--12',
-    'vads-u-margin-bottom--1',
-    'vads-u-margin-right--1',
-    'medium-screen:vads-l-col',
-  );
-
-  return (
-    <div className="vads-l-grid-container vads-u-padding--0">
-      <div className="vads-l-row">
-        {cards.map((card, index) => (
-          <div key={index} className={columnClassName}>
-            {card}
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
+import CardRow from './CardRow';
+import Card from './Card';
 
 export default function TopTasks() {
   return (
